Move classes into their own folder
