+++ functions/compare-build.js
@@ -1,4 +1,4 @@
-const SemVer = require('../semver')
+const SemVer = require('../classes/semver')
 
 module.exports = function compareBuild (a, b, loose) {
   const versionA = new SemVer(a, loose)
