+++ functions/compare.js
@@ -1,4 +1,4 @@
-const SemVer = require('../semver')
+const SemVer = require('../classes/semver')
 
 module.exports = function compare (a, b, loose) {
   return new SemVer(a, loose).compare(new SemVer(b, loose))
