+++ functions/minor.js
@@ -1,4 +1,4 @@
-const SemVer = require('../semver')
+const SemVer = require('../classes/semver')
 
 module.exports = function minor (a, loose) {
   return new SemVer(a, loose).minor
