+++ functions/patch.js
@@ -1,4 +1,4 @@
-const SemVer = require('../semver')
+const SemVer = require('../classes/semver')
 
 module.exports = function patch (a, loose) {
   return new SemVer(a, loose).patch
