+++ index.js
@@ -1,4 +1,5 @@
-const SemVer = require('./semver')
+const SemVer = require('./classes/semver')
+const { Comparator, Range, satisfies} = require('./classes/range')
 const { compareIdentifiers, rcompareIdentifiers } = require('./internal/identifiers')
 const { SEMVER_SPEC_VERSION } = require('./internal/constants')
 const { re, src, t } = require('./internal/re')
@@ -13,41 +14,40 @@ exports.SemVer = SemVer
 exports.compareIdentifiers = compareIdentifiers
 exports.rcompareIdentifiers = rcompareIdentifiers
 
-exports.parse = parse = require('./functions/parse')
-exports.valid = valid = require('./functions/valid')
-exports.clean = clean = require('./functions/clean')
-exports.inc = inc = require('./functions/inc')
-exports.diff = diff = require('./functions/diff')
-exports.major = major = require('./functions/major')
-exports.minor = minor = require('./functions/minor')
-exports.patch = patch = require('./functions/patch')
-exports.prerelease = prerelease = require('./functions/prerelease')
-exports.compare = compare = require('./functions/compare')
-exports.rcompare = rcompare = require('./functions/rcompare')
-exports.compareLoose = compareLoose = require('./functions/compare-loose')
-exports.compareBuild = compareBuild = require('./functions/compare-build')
-exports.sort = sort = require('./functions/sort')
-exports.rsort = rsort = require('./functions/rsort')
-exports.gt = gt = require('./functions/gt')
-exports.lt = lt = require('./functions/lt')
-exports.eq = eq = require('./functions/eq')
-exports.neq = neq = require('./functions/neq')
-exports.gte = gte = require('./functions/gte')
-exports.lte = lte = require('./functions/lte')
-exports.cmp = cmp = require('./functions/cmp')
-exports.coerce = coerce = require('./functions/coerce')
+exports.parse = require('./functions/parse')
+exports.valid = require('./functions/valid')
+exports.clean = require('./functions/clean')
+exports.inc = require('./functions/inc')
+exports.diff = require('./functions/diff')
+exports.major = require('./functions/major')
+exports.minor = require('./functions/minor')
+exports.patch = require('./functions/patch')
+exports.prerelease = require('./functions/prerelease')
+exports.compare = require('./functions/compare')
+exports.rcompare = require('./functions/rcompare')
+exports.compareLoose = require('./functions/compare-loose')
+exports.compareBuild = require('./functions/compare-build')
+exports.sort = require('./functions/sort')
+exports.rsort = require('./functions/rsort')
+exports.gt = require('./functions/gt')
+exports.lt = require('./functions/lt')
+exports.eq = require('./functions/eq')
+exports.neq = require('./functions/neq')
+exports.gte = require('./functions/gte')
+exports.lte = require('./functions/lte')
+exports.cmp = require('./functions/cmp')
+exports.coerce = require('./functions/coerce')
 
-const { Comparator, Range, satisfies} = require('./ranges')
 exports.Comparator = Comparator
 exports.Range = Range
 exports.satisfies = satisfies
 
-exports.toComparators = toComparators = require('./ranges/to-comparators')
-exports.maxSatisfying = maxSatisfying = require('./ranges/max-satisfying')
-exports.minSatisfying = minSatisfying = require('./ranges/min-satisfying')
-exports.minVersion = minVersion = require('./ranges/min-version')
-exports.validRange = validRange = require('./ranges/valid-range')
-exports.outside = outside = require('./ranges/outside')
-exports.gtr = gtr = require('./ranges/gtr')
-exports.ltr = ltr = require('./ranges/ltr')
-exports.intersects = intersects = require('./ranges/intersects')
+exports.toComparators = require('./ranges/to-comparators')
+exports.maxSatisfying = require('./ranges/max-satisfying')
+exports.minSatisfying = require('./ranges/min-satisfying')
+exports.minVersion = require('./ranges/min-version')
+exports.validRange = require('./ranges/valid-range')
+exports.outside = require('./ranges/outside')
+exports.gtr = require('./ranges/gtr')
+exports.ltr = require('./ranges/ltr')
+exports.intersects = require('./ranges/intersects')
