+++ ranges/max-satisfying.js
@@ -1,5 +1,5 @@
-const SemVer = require('../semver')
-const { Range } = require('./index')
+const SemVer = require('../classes/semver')
+const { Range } = require('../classes/range')
 
 module.exports = function maxSatisfying (versions, range, options) {
     let max = null
